requirements.txt
django
django-money
django-phonenumber-field